requirements.txt
streamlit==0.75.0
pandas
Pillow==8.1.0
